fix(tests): strengthen T14 asserts and consolidate shared fixtures
Must-fix (T14): Replace 5 weak/tautological assertions:
- test_warehouses.py: assert iter_paginated called with correct workspace
  UUID (replaces tautological `is not None` on result.output)
- test_snapshots.py rename: assert PATCH body contains new displayName
  (replaces `"displayName" in data` which passed with old name)
- test_snapshots.py delete/roll: remove always-true disjunct
  (keep only "deleted"/"rolled" signal, drop fixture-name fallback)
- test_snapshots.py --at: pin exact datetime(2024,3,15,12,0,0,UTC)
  (replaces `is not None` that couldn't catch wrong value)
- test_sql_pools.py since/until: pin exact naive datetime values
  (replaces `is not None` that couldn't catch wrong value;
   parse_iso_optional uses assume_utc=False so input stays naive)

Nit (cleanup): consolidate runner/cache_env/_real_http_client_cm
- Add tests/unit/cli/commands/conftest.py with all three shared fixtures
- Remove duplicate @pytest.fixture runner/cache_env from 16 test files
- Move _real_http_client_cm to conftest; import it in test_warehouses_respx.py
