+++ tests/unit/cli/commands/test_warehouses_respx.py
@@ -41,21 +41,16 @@ Each follows the same pattern established here.
 from __future__ import annotations
 
 import json
-import time
-from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
 from pathlib import Path
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import patch
 
 import httpx
-import pytest
 import respx
-from azure.core.credentials import AccessToken
 from click.testing import CliRunner
 
 from fabric_dw.cli._main import cli
-from fabric_dw.http_client import FabricHttpClient
 from tests.fixtures.api_payloads import WAREHOUSE_GET_PAYLOAD
+from tests.unit.cli.commands.conftest import _real_http_client_cm
 
 # ---------------------------------------------------------------------------
 # Constants matching fixtures
@@ -74,17 +69,6 @@ _WAREHOUSE_URL = f"{_BASE}/workspaces/{WS_GUID}/warehouses/{WH_GUID}"
 # ---------------------------------------------------------------------------
 
 
-def _fake_token() -> AccessToken:
-    return AccessToken(token="fake-bearer", expires_on=int(time.time()) + 3600)  # noqa: S106
-
-
-def _make_fake_credential() -> MagicMock:
-    """Return a mock AsyncTokenCredential that yields a long-lived fake token."""
-    cred = MagicMock()
-    cred.get_token = AsyncMock(return_value=_fake_token())
-    return cred
-
-
 # Generic items response needed by the Resolver GUID fast-path:
 # GET /workspaces/{ws}/items/{item} discovers the item type.
 _ITEMS_GENERIC_RESPONSE = {
@@ -97,33 +81,9 @@ _ITEMS_GENERIC_RESPONSE = {
 _WAREHOUSE_DETAIL = json.loads(WAREHOUSE_GET_PAYLOAD)
 
 
-@asynccontextmanager
-async def _real_http_client_cm(_ctx: object) -> AsyncIterator[FabricHttpClient]:
-    """Context manager that yields a *real* FabricHttpClient backed by a fake credential.
-
-    Use this as the replacement for ``build_http_client`` in respx-based tests.
-    The returned client is fully functional but its HTTP calls are intercepted by
-    whichever ``respx.mock`` context is active at call time.
-    """
-    cred = _make_fake_credential()
-    async with FabricHttpClient(credential=cred, rps=100) as client:
-        yield client
-
-
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
-
-
-@pytest.fixture
-def runner() -> CliRunner:
-    return CliRunner()
-
-
-@pytest.fixture
-def cache_env(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> Path:
-    monkeypatch.setenv("XDG_CACHE_HOME", str(tmp_path))
-    return tmp_path
 
 
 # ---------------------------------------------------------------------------
